Reorganized utils and moved models
